% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracing_objects.R
\name{Annotation.attributes}
\alias{Annotation.attributes}
\title{Annotation.attributes Object}
\usage{
Annotation.attributes()
}
\value{
Annotation.attributes object
}
\description{
Annotation.attributes Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A set of attributes on the annotation.
}
\seealso{
Other Annotation functions: \code{\link{Annotation}}
}
