% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracing_objects.R
\name{Annotation}
\alias{Annotation}
\title{Annotation Object}
\usage{
Annotation(Annotation.attributes = NULL, description = NULL,
  attributes = NULL)
}
\arguments{
\item{Annotation.attributes}{The \link{Annotation.attributes} object or list of objects}

\item{description}{A user-supplied message describing the event}

\item{attributes}{A set of attributes on the annotation}
}
\value{
Annotation object
}
\description{
Annotation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Text annotation with a set of attributes.
}
\seealso{
Other Annotation functions: \code{\link{Annotation.attributes}}
}
