% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracing_objects.R
\name{AttributeValue}
\alias{AttributeValue}
\title{AttributeValue Object}
\usage{
AttributeValue(intValue = NULL, stringValue = NULL, boolValue = NULL)
}
\arguments{
\item{intValue}{An integer value}

\item{stringValue}{A string value}

\item{boolValue}{A boolean value}
}
\value{
AttributeValue object
}
\description{
AttributeValue Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Allowed attribute values.
}
