% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracing_objects.R
\name{BatchUpdateSpansRequest}
\alias{BatchUpdateSpansRequest}
\title{BatchUpdateSpansRequest Object}
\usage{
BatchUpdateSpansRequest(BatchUpdateSpansRequest.spanUpdates = NULL,
  spanUpdates = NULL)
}
\arguments{
\item{BatchUpdateSpansRequest.spanUpdates}{The \link{BatchUpdateSpansRequest.spanUpdates} object or list of objects}

\item{spanUpdates}{A map from trace name to spans to be stored or updated}
}
\value{
BatchUpdateSpansRequest object
}
\description{
BatchUpdateSpansRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The request message for the `BatchUpdateSpans` method.
}
\seealso{
Other BatchUpdateSpansRequest functions: \code{\link{BatchUpdateSpansRequest.spanUpdates}},
  \code{\link{projects.traces.batchUpdate}}
}
