% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracing_objects.R
\name{BatchUpdateSpansRequest.spanUpdates}
\alias{BatchUpdateSpansRequest.spanUpdates}
\title{BatchUpdateSpansRequest.spanUpdates Object}
\usage{
BatchUpdateSpansRequest.spanUpdates()
}
\value{
BatchUpdateSpansRequest.spanUpdates object
}
\description{
BatchUpdateSpansRequest.spanUpdates Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A map from trace name to spans to be stored or updated.
}
\seealso{
Other BatchUpdateSpansRequest functions: \code{\link{BatchUpdateSpansRequest}},
  \code{\link{projects.traces.batchUpdate}}
}
