% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracing_objects.R
\name{Link}
\alias{Link}
\title{Link Object}
\usage{
Link(type = NULL, traceId = NULL, spanId = NULL)
}
\arguments{
\item{type}{The type of the link}

\item{traceId}{The trace identifier of the linked span}

\item{spanId}{The span identifier of the linked span}
}
\value{
Link object
}
\description{
Link Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Link one span with another which may be in a different Trace. Used (forexample) in batching operations, where a single batch handler processesmultiple requests from different traces.
}
