% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracing_objects.R
\name{ListSpansResponse}
\alias{ListSpansResponse}
\title{ListSpansResponse Object}
\usage{
ListSpansResponse(spans = NULL, nextPageToken = NULL)
}
\arguments{
\item{spans}{The requested spans if they are any in the specified trace}

\item{nextPageToken}{If defined, indicates that there are more spans that match the request}
}
\value{
ListSpansResponse object
}
\description{
ListSpansResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The response message for the 'ListSpans' method.
}
