% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracing_objects.R
\name{ListTracesResponse}
\alias{ListTracesResponse}
\title{ListTracesResponse Object}
\usage{
ListTracesResponse(traces = NULL, nextPageToken = NULL)
}
\arguments{
\item{traces}{List of trace records returned}

\item{nextPageToken}{If defined, indicates that there are more traces that match the request}
}
\value{
ListTracesResponse object
}
\description{
ListTracesResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The response message for the `ListTraces` method.
}
