% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracing_objects.R
\name{Module}
\alias{Module}
\title{Module Object}
\usage{
Module(buildId = NULL, module = NULL)
}
\arguments{
\item{buildId}{Build_id is a unique identifier for the module,}

\item{module}{E}
}
\value{
Module object
}
\description{
Module Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Binary module.
}
