% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracing_objects.R
\name{NetworkEvent}
\alias{NetworkEvent}
\title{NetworkEvent Object}
\usage{
NetworkEvent(kernelTime = NULL, type = NULL, messageId = NULL,
  messageSize = NULL)
}
\arguments{
\item{kernelTime}{If available, this is the kernel time:}

\item{type}{Type of a NetworkEvent}

\item{messageId}{Every message has an identifier, which must be different from all the}

\item{messageSize}{Number of bytes send/receive}
}
\value{
NetworkEvent object
}
\description{
NetworkEvent Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An event describing an RPC message sent/received on the network.
}
