% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracing_functions.R
\name{projects.traces.batchUpdate}
\alias{projects.traces.batchUpdate}
\title{Sends new spans to Stackdriver Trace or updates existing spans. If thename of a trace that you send matches that of an existing trace, any fieldsin the existing trace and its spans are overwritten by the provided values,and any new fields provided are merged with the existing trace data. If thename does not match, a new trace is created with given set of spans.}
\usage{
projects.traces.batchUpdate(BatchUpdateSpansRequest, parent)
}
\arguments{
\item{BatchUpdateSpansRequest}{The \link{BatchUpdateSpansRequest} object to pass to this method}

\item{parent}{ID of the Cloud project where the trace data is stored}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/trace.append
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/trace.append)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/trace}{Google Documentation}

Other BatchUpdateSpansRequest functions: \code{\link{BatchUpdateSpansRequest.spanUpdates}},
  \code{\link{BatchUpdateSpansRequest}}
}
