% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracing_functions.R
\name{projects.traces.list}
\alias{projects.traces.list}
\title{Returns of a list of traces that match the specified filter conditions.}
\usage{
projects.traces.list(parent, orderBy = NULL, filter = NULL,
  endTime = NULL, pageToken = NULL, startTime = NULL, pageSize = NULL)
}
\arguments{
\item{parent}{ID of the Cloud project where the trace data is stored}

\item{orderBy}{Field used to sort the returned traces}

\item{filter}{An optional filter for the request}

\item{endTime}{End of the time interval (inclusive) during which the trace data was}

\item{pageToken}{Token identifying the page of results to return}

\item{startTime}{Start of the time interval (inclusive) during which the trace data was}

\item{pageSize}{Maximum number of traces to return}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/trace.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/trace.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/trace}{Google Documentation}
}
