% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracing_functions.R
\name{projects.traces.listSpans}
\alias{projects.traces.listSpans}
\title{Returns a list of spans within a trace.}
\usage{
projects.traces.listSpans(name, pageToken = NULL)
}
\arguments{
\item{name}{ID of the span set where is 'projects/<project_id>/traces/<trace_id>'}

\item{pageToken}{Token identifying the page of results to return}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/trace.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/trace.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/trace}{Google Documentation}
}
