% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracing_objects.R
\name{Span.attributes}
\alias{Span.attributes}
\title{Span.attributes Object}
\usage{
Span.attributes()
}
\value{
Span.attributes object
}
\description{
Span.attributes Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Properties of a span. Attributes at the span level.E.g.'/instance_id': 'my-instance''/zone': 'us-central1-a''/grpc/peer_address': 'ip:port' (dns, etc.)'/grpc/deadline': 'Duration''/http/user_agent''/http/request_bytes': 300'/http/response_bytes': 1200'/http/url': google.com/apis'/pid''abc.com/myattribute': 'my attribute value'Maximum length for attribute key is 128 characters, for string attributevalue is 2K characters.
}
\seealso{
Other Span functions: \code{\link{Span}}
}
