% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracing_objects.R
\name{Span}
\alias{Span}
\title{Span Object}
\usage{
Span(Span.attributes = NULL, localStartTime = NULL,
  hasRemoteParent = NULL, timeEvents = NULL, parentId = NULL,
  localEndTime = NULL, status = NULL, name = NULL, stackTrace = NULL,
  links = NULL, attributes = NULL, id = NULL)
}
\arguments{
\item{Span.attributes}{The \link{Span.attributes} object or list of objects}

\item{localStartTime}{Local machine clock time from the UNIX epoch,}

\item{hasRemoteParent}{True if this Span has a remote parent (is an RPC server Span)}

\item{timeEvents}{A collection of time-stamped events}

\item{parentId}{ID of parent span}

\item{localEndTime}{Local machine clock time from the UNIX epoch,}

\item{status}{The final status of the Span}

\item{name}{Name of the span}

\item{stackTrace}{Stack trace captured at the start of the span}

\item{links}{A collection of links}

\item{attributes}{Properties of a span}

\item{id}{Identifier for the span}
}
\value{
Span object
}
\description{
Span Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A span represents a single operation within a trace. Spans can be nestedand form a trace tree. Often, a trace contains a root span that describes theend-to-end latency and, optionally, one or more subspans forits sub-operations. Spans do not need to be contiguous. There may be gapsbetween spans in a trace.
}
\seealso{
Other Span functions: \code{\link{Span.attributes}}
}
