% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracing_objects.R
\name{SpanUpdates}
\alias{SpanUpdates}
\title{SpanUpdates Object}
\usage{
SpanUpdates(spans = NULL)
}
\arguments{
\item{spans}{A collection of spans}
}
\value{
SpanUpdates object
}
\description{
SpanUpdates Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Collection of spans to update.
}
