% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracing_objects.R
\name{StackFrame}
\alias{StackFrame}
\title{StackFrame Object}
\usage{
StackFrame(loadModule = NULL, columnNumber = NULL, fileName = NULL,
  sourceVersion = NULL, originalFunctionName = NULL, functionName = NULL,
  lineNumber = NULL)
}
\arguments{
\item{loadModule}{Binary module the code is loaded from}

\item{columnNumber}{Column number is important in JavaScript(anonymous functions),}

\item{fileName}{File name of the frame}

\item{sourceVersion}{source_version is deployment specific}

\item{originalFunctionName}{Used when function name is ‘mangled’}

\item{functionName}{Fully qualified names which uniquely identify function/method/etc}

\item{lineNumber}{Line number of the frame}
}
\value{
StackFrame object
}
\description{
StackFrame Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Presents a single stack frame in a stack trace.
}
