% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracing_objects.R
\name{StackTrace}
\alias{StackTrace}
\title{StackTrace Object}
\usage{
StackTrace(stackFrame = NULL, stackTraceHashId = NULL)
}
\arguments{
\item{stackFrame}{Stack frames of this stack trace}

\item{stackTraceHashId}{User can choose to use their own hash function to hash large attributes to}
}
\value{
StackTrace object
}
\description{
StackTrace Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
StackTrace collected in a trace.
}
