% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracing_objects.R
\name{TimeEvent}
\alias{TimeEvent}
\title{TimeEvent Object}
\usage{
TimeEvent(annotation = NULL, localTime = NULL, networkEvent = NULL)
}
\arguments{
\item{annotation}{Optional field for user supplied <string, AttributeValue> map}

\item{localTime}{The local machine absolute timestamp when this event happened}

\item{networkEvent}{Optional field that can be used only for network events}
}
\value{
TimeEvent object
}
\description{
TimeEvent Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A time-stamped annotation in the Span.
}
