% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracing_objects.R
\name{Trace}
\alias{Trace}
\title{Google Tracing API Objects 
Send and retrieve trace data from Google Stackdriver Trace.}
\usage{
Trace(name = NULL)
}
\arguments{
\item{name}{ID of the trace which is 'projects/<project_id>/traces/<trace_id>'}
}
\value{
Trace object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2017-03-05 20:20:46
filename: /Users/mark/dev/R/autoGoogleAPI/googletracingv1.auto/R/tracing_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
Trace Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A trace describes how long it takes for an application to perform someoperations. It consists of a set of spans, each of which contains detailsabout an operation with time information and operation details.
}
