% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracing_functions.R
\docType{package}
\name{tracing_googleAuthR}
\alias{tracing_googleAuthR}
\alias{tracing_googleAuthR-package}
\title{Google Tracing API
Send and retrieve trace data from Google Stackdriver Trace.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2017-03-05 20:20:46
filename: /Users/mark/dev/R/autoGoogleAPI/googletracingv1.auto/R/tracing_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/trace.readonly
\item https://www.googleapis.com/auth/trace.append
\item https://www.googleapis.com/auth/cloud-platform
}
}
